Return BankError to the Bank
