--- src/bank_error.rs
@@ -1,47 +1,0 @@
-use solana_sdk::native_program::ProgramError;
-use std::result;
-/// Reasons a transaction might be rejected.
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub enum BankError {
-    /// This Pubkey is being processed in another transaction
-    AccountInUse,
-
-    /// Attempt to debit from `Pubkey`, but no found no record of a prior credit.
-    AccountNotFound,
-
-    /// The from `Pubkey` does not have sufficient balance to pay the fee to schedule the transaction
-    InsufficientFundsForFee,
-
-    /// The bank has seen `Signature` before. This can occur under normal operation
-    /// when a UDP packet is duplicated, as a user error from a client not updating
-    /// its `last_id`, or as a double-spend attack.
-    DuplicateSignature,
-
-    /// The bank has not seen the given `last_id` or the transaction is too old and
-    /// the `last_id` has been discarded.
-    LastIdNotFound,
-
-    /// The bank has not seen a transaction with the given `Signature` or the transaction is
-    /// too old and has been discarded.
-    SignatureNotFound,
-
-    /// A transaction with this signature has been received but not yet executed
-    SignatureReserved,
-
-    /// Proof of History verification failed.
-    LedgerVerificationFailed,
-
-    /// The program returned an error
-    ProgramError(u8, ProgramError),
-
-    /// Recoding into PoH failed
-    RecordFailure,
-
-    /// Loader call chain too deep
-    CallChainTooDeep,
-
-    /// Transaction has a fee but has no signature present
-    MissingSignatureForFee,
-}
-
-pub type Result<T> = result::Result<T, BankError>;
